refactor: moved templates to  internal package
