+++ internal/templates/templates.go
@@ -1,19 +1,19 @@
-package functioncheck
+package templates
 
 const (
-	allowedMacroTemplate = `
+	AllowedMacroTemplate = `
 #[cfg(not(feature = "allowed_%s"))]
 #[macro_export]
 macro_rules! %s {
 	($($arg:tt)*) => {{}}
 }
 `
-	allowedFunctionTemplate = `
+	AllowedFunctionTemplate = `
 #[cfg(not(feature = "allowed_%s"))]
 pub fn %s() {}
 `
-	allowedItemsLibHeader = "pub mod %s { "
-	cargoTomlTemplate     = `[package]
+	AllowedItemsLibHeader = "pub mod %s { "
+	CargoTomlTemplate     = `[package]
 name = "%s"
 version = "0.1.0"
 edition = "2021"
@@ -27,17 +27,17 @@ path = "src/%s/temp.rs"
 
 [workspace]
 `
-	allowedItemsCargoToml = `[package]
+	AllowedItemsCargoToml = `[package]
 name = "allowedfunctions"
 version = "0.1.0"
 edition = "2021"
 `
-	studentCodePrefix = `#![no_std]
+	StudentCodePrefix = `#![no_std]
 #[macro_use]
 extern crate allowedfunctions;
 use allowedfunctions::%s::*;
 `
-	dummyMain = `
+	DummyMain = `
 fn main() {
 	%s;
 }`
